Refactoring of transit route processing.
