+++ drape_frontend/drape_engine.cpp
@@ -241,7 +241,6 @@ void DrapeEngine::UpdateUserMarksGroup(MarkGroupID groupId, UserMarksProvider * 
       renderInfo->m_pivot = mark->GetPivot();
       renderInfo->m_pixelOffset = mark->GetPixelOffset();
       renderInfo->m_titleDecl = mark->GetTitleDecl();
-      renderInfo->m_symbolSizes = mark->GetSymbolSizes();
       renderInfo->m_symbolNames = mark->GetSymbolNames();
       renderInfo->m_coloredSymbols = mark->GetColoredSymbols();
       renderInfo->m_hasSymbolPriority = mark->HasSymbolPriority();
