+++ map/routing_manager.cpp
@@ -17,9 +17,7 @@
 
 #include "routing_common/num_mwm_id.hpp"
 
-#include "drape_frontend/color_constants.hpp"
 #include "drape_frontend/drape_engine.hpp"
-#include "drape_frontend/visual_params.hpp"
 
 #include "indexer/map_style_reader.hpp"
 #include "indexer/scales.hpp"
@@ -57,32 +55,10 @@ namespace
 char const kRouterTypeKey[] = "router";
 
 double const kRouteScaleMultiplier = 1.5;
-float const kStopMarkerScale = 2.2f;
-float const kTransferMarkerScale = 4.0f;
 
 string const kRoutePointsFile = "route_points.dat";
 
 uint32_t constexpr kInvalidTransactionId = 0;
-
-map<TransitType, string> const kTransitSymbols = {
-    { TransitType::Subway, "transit_subway" },
-    { TransitType::LightRail, "transit_light_rail" },
-    { TransitType::Monorail, "transit_monorail" },
-    { TransitType::Train, "transit_train" }
-};
-
-TransitType GetTransitType(string const & type)
-{
-  if (type == "subway")
-    return TransitType::Subway;
-  if (type == "light_rail")
-    return TransitType::LightRail;
-  if (type == "monorail")
-    return TransitType::Monorail;
-
-  ASSERT_EQUAL(type, "train", ());
-  return TransitType::Train;
-}
 
 void FillTurnsDistancesForRendering(vector<RouteSegment> const & segments,
                                     double baseDistance, vector<double> & turns)
@@ -112,493 +88,6 @@ void FillTrafficForRendering(vector<RouteSegment> const & segments,
   traffic.reserve(segments.size());
   for (auto const & s : segments)
     traffic.push_back(s.GetTraffic());
-}
-
-struct TransitTitle
-{
-  TransitTitle() = default;
-  TransitTitle(string const & text, df::ColorConstant const & color) : m_text(text), m_color(color) {}
-
-  string m_text;
-  df::ColorConstant m_color;
-};
-
-struct TransitMarkInfo
-{
-  enum class Type
-  {
-    Stop,
-    KeyStop,
-    Transfer,
-    Gate
-  };
-  Type m_type = Type::Stop;
-  m2::PointD m_point;
-  vector<TransitTitle> m_titles;
-  std::string m_symbolName;
-  df::ColorConstant m_color;
-  FeatureID m_featureId;
-};
-
-using GetMwmIdFn = function<MwmSet::MwmId (routing::NumMwmId numMwmId)>;
-void CollectTransitDisplayInfo(vector<RouteSegment> const & segments, GetMwmIdFn const & getMwmIdFn,
-                               TransitDisplayInfos & transitDisplayInfos)
-{
-  for (auto const & s : segments)
-  {
-    if (!s.HasTransitInfo())
-      continue;
-
-    auto const mwmId = getMwmIdFn(s.GetSegment().GetMwmId());
-
-    auto & mwmTransit = transitDisplayInfos[mwmId];
-    if (mwmTransit == nullptr)
-      mwmTransit = my::make_unique<TransitDisplayInfo>();
-
-    TransitInfo const & transitInfo = s.GetTransitInfo();
-    switch (transitInfo.GetType())
-    {
-      case TransitInfo::Type::Edge:
-      {
-        auto const & edge = transitInfo.GetEdge();
-
-        mwmTransit->m_stops[edge.m_stop1Id] = {};
-        mwmTransit->m_stops[edge.m_stop2Id] = {};
-        mwmTransit->m_lines[edge.m_lineId] = {};
-        for (auto const &shapeId : edge.m_shapeIds)
-          mwmTransit->m_shapes[shapeId] = {};
-        break;
-      }
-      case TransitInfo::Type::Transfer:
-      {
-        auto const & transfer = transitInfo.GetTransfer();
-        mwmTransit->m_stops[transfer.m_stop1Id] = {};
-        mwmTransit->m_stops[transfer.m_stop2Id] = {};
-        break;
-      }
-      case TransitInfo::Type::Gate:
-      {
-        auto const & gate = transitInfo.GetGate();
-        if (gate.m_featureId != transit::kInvalidFeatureId)
-        {
-          auto const featureId = FeatureID(mwmId, gate.m_featureId);
-          TransitFeatureInfo featureInfo;
-          featureInfo.m_isGate = true;
-          mwmTransit->m_features[featureId] = featureInfo;
-        }
-        break;
-      }
-    }
-  }
-}
-
-void AddTransitGateSegment(m2::PointD const & destPoint, df::ColorConstant const & color, df::Subroute & subroute)
-{
-  ASSERT_GREATER(subroute.m_polyline.GetSize(), 0, ());
-  df::SubrouteStyle style(color, df::RoutePattern(4.0, 2.0));
-  style.m_startIndex = subroute.m_polyline.GetSize() - 1;
-  auto const vec = destPoint - subroute.m_polyline.Back();
-  subroute.m_polyline.Add(destPoint);
-  style.m_endIndex = subroute.m_polyline.GetSize() - 1;
-  subroute.AddStyle(style);
-}
-
-void AddTransitPedestrianSegment(m2::PointD const & destPoint, df::Subroute & subroute)
-{
-  ASSERT_GREATER(subroute.m_polyline.GetSize(), 0, ());
-  df::SubrouteStyle style(df::kRoutePedestrian, df::RoutePattern(4.0, 2.0));
-  style.m_startIndex = subroute.m_polyline.GetSize() - 1;
-  subroute.m_polyline.Add(destPoint);
-  style.m_endIndex = subroute.m_polyline.GetSize() - 1;
-  subroute.AddStyle(style);
-}
-
-void AddTransitShapes(std::vector<transit::ShapeId> const & shapeIds, TransitShapesInfo const & shapes,
-                      df::ColorConstant const & color, bool isInverted, df::Subroute & subroute)
-{
-  ASSERT_GREATER(subroute.m_polyline.GetSize(), 0, ());
-  df::SubrouteStyle style(color);
-  style.m_startIndex = subroute.m_polyline.GetSize() - 1;
-
-  for (auto it = shapeIds.crbegin(); it != shapeIds.crend(); ++it)
-  {
-    auto const & shapePolyline = shapes.at(*it).GetPolyline();
-    if (isInverted)
-      subroute.m_polyline.Append(shapePolyline.crbegin(), shapePolyline.crend());
-    else
-      subroute.m_polyline.Append(shapePolyline.cbegin(), shapePolyline.cend());
-  }
-
-  style.m_endIndex = subroute.m_polyline.GetSize() - 1;
-  subroute.AddStyle(style);
-}
-
-uint32_t ColorToARGB(df::ColorConstant const & colorConstant)
-{
-  auto const color = df::GetColorConstant(colorConstant);
-  return color.GetAlpha() << 24 | color.GetRed() << 16 | color.GetGreen() << 8 | color.GetBlue();
-}
-
-void FillTransitStyleForRendering(vector<RouteSegment> const & segments, TransitReadManager & transitReadManager,
-                                  GetMwmIdFn const & getMwmIdFn,
-                                  RoutingManager::Callbacks::GetStringsBundleFn const & getStringsBundleFn,
-                                  df::Subroute & subroute, vector<TransitMarkInfo> & transitMarks,
-                                  TransitRouteInfo & routeInfo)
-{
-  TransitDisplayInfos transitDisplayInfos;
-  CollectTransitDisplayInfo(segments, getMwmIdFn, transitDisplayInfos);
-
-  // Read transit display info.
-  if (!transitReadManager.GetTransitDisplayInfo(transitDisplayInfos))
-    return;
-
-  subroute.m_styleType = df::SubrouteStyleType::Multiple;
-  subroute.m_style.clear();
-  subroute.m_style.reserve(segments.size());
-
-  df::ColorConstant lastColor;
-  m2::PointD lastDir;
-
-  df::SubrouteMarker marker;
-  TransitMarkInfo transitMarkInfo;
-  TransitType transitType = TransitType::Pedestrian;
-
-  double prevDistance = routeInfo.m_totalDistInMeters;
-  double prevTime = routeInfo.m_totalTimeInSec;
-
-  bool pendingEntrance = false;
-
-  for (auto const & s : segments)
-  {
-    auto const time = static_cast<int>(ceil(s.GetTimeFromBeginningSec() - prevTime));
-    auto const distance = s.GetDistFromBeginningMeters() - prevDistance;
-    prevDistance = s.GetDistFromBeginningMeters();
-    prevTime = s.GetTimeFromBeginningSec();
-
-    if (!s.HasTransitInfo())
-    {
-      routeInfo.AddStep(TransitStepInfo(TransitType::Pedestrian, distance, time));
-
-      AddTransitPedestrianSegment(s.GetJunction().GetPoint(), subroute);
-      lastColor = "";
-      transitType = TransitType::Pedestrian;
-      continue;
-    }
-
-    auto const mwmId = getMwmIdFn(s.GetSegment().GetMwmId());
-
-    auto const & transitInfo = s.GetTransitInfo();
-    auto const & displayInfo = *transitDisplayInfos.at(mwmId).get();
-
-    if (transitInfo.GetType() == TransitInfo::Type::Edge)
-    {
-      auto const & edge = transitInfo.GetEdge();
-
-      auto const & line = displayInfo.m_lines.at(edge.m_lineId);
-      auto const currentColor = df::GetTransitColorName(line.GetColor());
-      transitType = GetTransitType(line.GetType());
-
-      routeInfo.AddStep(TransitStepInfo(transitType, distance, time,
-                                        line.GetNumber(), ColorToARGB(currentColor)));
-
-      auto const & stop1 = displayInfo.m_stops.at(edge.m_stop1Id);
-      auto const & stop2 = displayInfo.m_stops.at(edge.m_stop2Id);
-      bool const isTransfer1 = stop1.GetTransferId() != transit::kInvalidTransferId;
-      bool const isTransfer2 = stop2.GetTransferId() != transit::kInvalidTransferId;
-
-      marker.m_distance = prevDistance;
-      marker.m_scale = kStopMarkerScale;
-      marker.m_innerColor = currentColor;
-      if (isTransfer1)
-      {
-        auto const & transferData = displayInfo.m_transfers.at(stop1.GetTransferId());
-        marker.m_position = transferData.GetPoint();
-      }
-      else
-      {
-        marker.m_position = stop1.GetPoint();
-      }
-      transitMarkInfo.m_point = marker.m_position;
-
-      if (pendingEntrance)
-      {
-        transitMarkInfo.m_type = TransitMarkInfo::Type::KeyStop;
-        transitMarkInfo.m_symbolName = kTransitSymbols.at(transitType);
-        transitMarkInfo.m_color = currentColor;
-        AddTransitGateSegment(marker.m_position, currentColor, subroute);
-        pendingEntrance = false;
-      }
-
-      auto const id1 = isTransfer1 ? stop1.GetTransferId() : stop1.GetId();
-      auto const id2 = isTransfer2 ? stop2.GetTransferId() : stop2.GetId();
-      bool const isInverted = id1 > id2;
-
-      AddTransitShapes(edge.m_shapeIds, displayInfo.m_shapes, currentColor, isInverted, subroute);
-
-      ASSERT_GREATER(subroute.m_polyline.GetSize(), 1, ());
-      auto const & p1 = *(subroute.m_polyline.End() - 2);
-      auto const & p2 = *(subroute.m_polyline.End() - 1);
-      m2::PointD currentDir = (p2 - p1).Normalize();
-
-      if (lastColor != currentColor)
-      {
-        if (!lastColor.empty())
-        {
-          marker.m_scale = kTransferMarkerScale;
-          transitMarkInfo.m_type = TransitMarkInfo::Type::Transfer;
-        }
-        marker.m_colors.push_back(currentColor);
-
-        if (stop1.GetFeatureId() != transit::kInvalidFeatureId)
-        {
-          auto const fid = FeatureID(mwmId, stop1.GetFeatureId());
-          transitMarkInfo.m_featureId = fid;
-          transitMarkInfo.m_titles.emplace_back(displayInfo.m_features.at(fid).m_title,
-                                                df::GetTransitTextColorName(line.GetColor()));
-        }
-      }
-      lastColor = currentColor;
-
-      if (marker.m_colors.size() > 1)
-      {
-        marker.m_innerColor = df::kTransitOutlineColor;
-        marker.m_up = (currentDir - lastDir).Normalize();
-        if (m2::CrossProduct(marker.m_up, -lastDir) < 0)
-          marker.m_up = -marker.m_up;
-      }
-
-      subroute.m_markers.push_back(marker);
-      marker = df::SubrouteMarker();
-
-      transitMarks.push_back(transitMarkInfo);
-      transitMarkInfo = TransitMarkInfo();
-
-      lastDir = currentDir;
-
-      marker.m_distance = s.GetDistFromBeginningMeters();
-      marker.m_scale = kStopMarkerScale;
-      marker.m_innerColor = currentColor;
-      marker.m_colors.push_back(currentColor);
-
-      if (isTransfer2)
-      {
-        auto const & transferData = displayInfo.m_transfers.at(stop2.GetTransferId());
-        marker.m_position = transferData.GetPoint();
-      }
-      else
-      {
-        marker.m_position = stop2.GetPoint();
-      }
-
-      transitMarkInfo.m_point = marker.m_position;
-      if (stop2.GetFeatureId() != transit::kInvalidFeatureId)
-      {
-        auto const fid = FeatureID(mwmId, stop2.GetFeatureId());
-        transitMarkInfo.m_featureId = fid;
-        transitMarkInfo.m_titles.push_back(TransitTitle(displayInfo.m_features.at(fid).m_title,
-                                                        df::GetTransitTextColorName(line.GetColor())));
-      }
-    }
-    else if (transitInfo.GetType() == TransitInfo::Type::Gate)
-    {
-      auto const & gate = transitInfo.GetGate();
-      if (!lastColor.empty())
-      {
-        routeInfo.AddStep(TransitStepInfo(TransitType::Pedestrian, distance, time));
-
-        AddTransitGateSegment(s.GetJunction().GetPoint(), lastColor, subroute);
-
-        subroute.m_markers.push_back(marker);
-        marker = df::SubrouteMarker();
-
-        transitMarkInfo.m_type = TransitMarkInfo::Type::KeyStop;
-        transitMarkInfo.m_symbolName = kTransitSymbols.at(transitType);
-        transitMarkInfo.m_color = lastColor;
-        transitMarks.push_back(transitMarkInfo);
-        transitMarkInfo = TransitMarkInfo();
-      }
-      else
-      {
-        pendingEntrance = true;
-      }
-
-      auto gateMarkInfo = TransitMarkInfo();
-      gateMarkInfo.m_point = pendingEntrance ? subroute.m_polyline.Back() : s.GetJunction().GetPoint();
-      gateMarkInfo.m_type = TransitMarkInfo::Type::Gate;
-      if (gate.m_featureId != transit::kInvalidFeatureId)
-      {
-        auto const fid = FeatureID(mwmId, gate.m_featureId);
-        auto const & featureInfo = displayInfo.m_features.at(fid);
-        auto symbolName = featureInfo.m_gateSymbolName;
-        if (strings::EndsWith(symbolName, "-s") ||
-            strings::EndsWith(symbolName, "-m") ||
-            strings::EndsWith(symbolName, "-l"))
-        {
-          symbolName = symbolName.substr(0, symbolName.length() - 2);
-        }
-
-        gateMarkInfo.m_featureId = fid;
-        gateMarkInfo.m_symbolName = symbolName;
-        auto const title = getStringsBundleFn().GetString(pendingEntrance ? "entrance" : "exit");
-        gateMarkInfo.m_titles.push_back(TransitTitle(title, df::GetTransitTextColorName("default")));
-      }
-
-      transitMarks.push_back(gateMarkInfo);
-    }
-  }
-
-  routeInfo.m_totalDistInMeters = prevDistance;
-  routeInfo.m_totalTimeInSec = static_cast<int>(ceil(prevTime));
-}
-
-vector<m2::PointF> GetTransitMarkerSizes(float markerScale)
-{
-  vector<m2::PointF> markerSizes;
-  markerSizes.reserve(df::kRouteHalfWidthInPixelTransit.size());
-  for (auto const halfWidth : df::kRouteHalfWidthInPixelTransit)
-  {
-    float const d = 2 * halfWidth * markerScale;
-    markerSizes.push_back(m2::PointF(d, d));
-  }
-  return markerSizes;
-}
-
-void CreateTransitMarks(vector<TransitMarkInfo> const & transitMarks, std::map<std::string, m2::PointF> symbolSizes,
-                        UserMarksController & marksController)
-{
-  static vector<m2::PointF> const kTransferMarkerSizes = GetTransitMarkerSizes(kTransferMarkerScale);
-  static vector<m2::PointF> const kStopMarkerSizes = GetTransitMarkerSizes(kStopMarkerScale);
-
-  static const int kSmallIconZoom = 1;
-  static const int kMediumIconZoom = 10;
-  static const int kLargeIconZoom = 15;
-  static const int kMinStopTitleZoom = 14;
-  static const float kGateBgScale = 1.2f;
-  static const int kTransferTitleOffset = 1;
-  static const int kStopTitleOffset = 0;
-  static const int kGateTitleOffset = 0;
-
-  auto const vs = df::VisualParams::Instance().GetVisualScale();
-  for (size_t i = 0; i < transitMarks.size(); ++i)
-  {
-    auto const &mark = transitMarks[i];
-
-    auto userMark = marksController.CreateUserMark(mark.m_point);
-    ASSERT(dynamic_cast<TransitMark *>(userMark) != nullptr, ());
-    auto transitMark = static_cast<TransitMark *>(userMark);
-    dp::TitleDecl titleDecl;
-
-    transitMark->SetFeatureId(mark.m_featureId);
-
-    if (mark.m_type == TransitMarkInfo::Type::Gate)
-    {
-      if (!mark.m_titles.empty())
-      {
-        TransitMark::GetDefaultTransitTitle(titleDecl);
-        titleDecl.m_primaryText = mark.m_titles.front().m_text;
-        titleDecl.m_primaryOffset.y = kGateTitleOffset;
-        titleDecl.m_anchor = dp::Anchor::Top;
-        titleDecl.m_primaryOptional = true;
-        transitMark->AddTitle(titleDecl);
-      }
-      df::UserPointMark::SymbolNameZoomInfo symbolNames;
-      symbolNames[kSmallIconZoom] = mark.m_symbolName + "-s";
-      symbolNames[kMediumIconZoom] = mark.m_symbolName + "-m";
-      symbolNames[kLargeIconZoom] = mark.m_symbolName + "-l";
-      transitMark->SetSymbolNames(symbolNames);
-      transitMark->SetPriority(UserMark::Priority::Transit_Gate);
-    }
-    else if (mark.m_type == TransitMarkInfo::Type::Transfer)
-    {
-      if (mark.m_titles.size() > 1)
-      {
-        TransitMark::GetDefaultTransitTitle(titleDecl);
-        titleDecl.m_primaryText = mark.m_titles.front().m_text;
-        titleDecl.m_primaryTextFont.m_color = df::GetColorConstant(mark.m_titles.front().m_color);
-        titleDecl.m_primaryOffset.y = -kTransferTitleOffset;
-        titleDecl.m_anchor = dp::Anchor::Bottom;
-        titleDecl.m_primaryOptional = true;
-        transitMark->AddTitle(titleDecl);
-
-        titleDecl.m_primaryText = mark.m_titles.back().m_text;
-        titleDecl.m_primaryTextFont.m_color = df::GetColorConstant(mark.m_titles.back().m_color);
-        titleDecl.m_primaryOffset.y = kTransferTitleOffset;
-        titleDecl.m_anchor = dp::Anchor::Top;
-        titleDecl.m_primaryOptional = true;
-        transitMark->AddTitle(titleDecl);
-      }
-      df::UserPointMark::ColoredSymbolZoomInfo coloredSymbol;
-      for (size_t sizeIndex = 0; sizeIndex < kTransferMarkerSizes.size(); ++sizeIndex)
-      {
-        auto const zoomLevel = sizeIndex + 1;
-        auto const & sz = kTransferMarkerSizes[sizeIndex];
-        df::ColoredSymbolViewParams params;
-        params.m_radiusInPixels = static_cast<float>(max(sz.x, sz.y) * vs / 2.0);
-        params.m_color = dp::Color::Transparent();
-        if (coloredSymbol.empty() || coloredSymbol.rbegin()->second.m_radiusInPixels != params.m_radiusInPixels)
-          coloredSymbol.insert(make_pair(zoomLevel, params));
-      }
-      transitMark->SetColoredSymbols(coloredSymbol);
-      transitMark->SetPriority(UserMark::Priority::Transit_Transfer);
-    }
-    else
-    {
-      if (!mark.m_titles.empty())
-      {
-        TransitMark::GetDefaultTransitTitle(titleDecl);
-        titleDecl.m_primaryText = mark.m_titles.front().m_text;
-        titleDecl.m_primaryTextFont.m_color = df::GetColorConstant(mark.m_titles.front().m_color);
-        titleDecl.m_primaryOffset.y = kStopTitleOffset;
-        titleDecl.m_anchor = dp::Anchor::Top;
-        titleDecl.m_primaryOptional = true;
-        transitMark->AddTitle(titleDecl);
-      }
-      if (mark.m_type == TransitMarkInfo::Type::KeyStop)
-      {
-        df::UserPointMark::SymbolNameZoomInfo symbolNames;
-        symbolNames[kSmallIconZoom] = mark.m_symbolName + "-s";
-        symbolNames[kMediumIconZoom] = mark.m_symbolName + "-m";
-        symbolNames[kLargeIconZoom] = mark.m_symbolName + "-l";
-        transitMark->SetSymbolNames(symbolNames);
-
-        df::UserPointMark::ColoredSymbolZoomInfo coloredSymbol;
-        df::ColoredSymbolViewParams params;
-        params.m_color = df::GetColorConstant(mark.m_color);
-
-        auto sz = symbolSizes[symbolNames[kSmallIconZoom]];
-        params.m_radiusInPixels = static_cast<float>(max(sz.x, sz.y) * kGateBgScale / 2.0);
-        coloredSymbol[kSmallIconZoom] = params;
-
-        sz = symbolSizes[symbolNames[kMediumIconZoom]];
-        params.m_radiusInPixels = static_cast<float>(max(sz.x, sz.y) * kGateBgScale / 2.0);
-        coloredSymbol[kMediumIconZoom] = params;
-
-        sz = symbolSizes[symbolNames[kLargeIconZoom]];
-        params.m_radiusInPixels = static_cast<float>(max(sz.x, sz.y) * kGateBgScale / 2.0);
-        coloredSymbol[kLargeIconZoom] = params;
-
-        transitMark->SetColoredSymbols(coloredSymbol);
-        transitMark->SetPriority(UserMark::Priority::Transit_KeyStop);
-      }
-      else
-      {
-        df::UserPointMark::ColoredSymbolZoomInfo coloredSymbol;
-        for (size_t sizeIndex = 0; sizeIndex < kStopMarkerSizes.size(); ++sizeIndex)
-        {
-          auto const zoomLevel = sizeIndex + 1;
-          auto const & sz = kStopMarkerSizes[sizeIndex];
-          df::ColoredSymbolViewParams params;
-          params.m_radiusInPixels = static_cast<float>(max(sz.x, sz.y) * vs / 2.0);
-          params.m_color = dp::Color::Transparent();
-          if (coloredSymbol.empty() || coloredSymbol.rbegin()->second.m_radiusInPixels != params.m_radiusInPixels)
-            coloredSymbol.insert(make_pair(zoomLevel, params));
-        }
-        transitMark->SetColoredSymbols(coloredSymbol);
-        transitMark->SetPriority(UserMark::Priority::Transit_Stop);
-        transitMark->SetMinTitleZoom(kMinStopTitleZoom);
-      }
-    }
-  }
 }
 
 RouteMarkData GetLastPassedPoint(BookmarkManager * bmManager, vector<RouteMarkData> const & points)
@@ -733,59 +222,6 @@ namespace marketing
 {
 char const * const kRoutingCalculatingRoute = "Routing_CalculatingRoute";
 }  // namespace marketing
-
-TransitStepInfo::TransitStepInfo(TransitType type, double distance, int time,
-                                 std::string const & number, uint32_t color,
-                                 int intermediateIndex)
-  : m_type(type)
-  , m_distanceInMeters(distance)
-  , m_timeInSec(time)
-  , m_number(number)
-  , m_colorARGB(color)
-  , m_intermediateIndex(intermediateIndex)
-{}
-
-bool TransitStepInfo::IsEqualType(TransitStepInfo const & ts) const
-{
-  if (m_type != ts.m_type)
-    return false;
-
-  if (m_type != TransitType::Pedestrian && m_type != TransitType::IntermediatePoint)
-    return m_number == ts.m_number && m_colorARGB == ts.m_colorARGB;
-  return true;
-}
-
-void TransitRouteInfo::AddStep(TransitStepInfo const & step)
-{
-  if (!m_steps.empty() && m_steps.back().IsEqualType(step))
-  {
-    m_steps.back().m_distanceInMeters += step.m_distanceInMeters;
-    m_steps.back().m_timeInSec += step.m_timeInSec;
-  }
-  else
-  {
-    m_steps.push_back(step);
-  }
-
-  if (step.m_type == TransitType::Pedestrian)
-  {
-    m_totalPedestrianDistInMeters += step.m_distanceInMeters;
-    m_totalPedestrianTimeInSec += step.m_timeInSec;
-  }
-}
-
-void TransitRouteInfo::UpdateDistanceStrings()
-{
-  if (m_steps.empty())
-    return;
-  for (auto & step : m_steps)
-  {
-    FormatDistance(step.m_distanceInMeters, step.m_distanceStr, step.m_distanceUnitsSuffix);
-  }
-  FormatDistance(m_totalDistInMeters, m_totalDistanceStr, m_totalDistanceUnitsSuffix);
-  FormatDistance(m_totalPedestrianDistInMeters, m_totalPedestrianDistanceStr,
-                 m_totalPedestrianUnitsSuffix);
-}
 
 RoutingManager::RoutingManager(Callbacks && callbacks, Delegate & delegate)
   : m_callbacks(move(callbacks))
@@ -1008,10 +444,24 @@ void RoutingManager::InsertRoute(Route const & route)
   // TODO: Now we always update whole route, so we need to remove previous one.
   RemoveRoute(false /* deactivateFollowing */);
 
+  std::unique_ptr<TransitRouteDisplay> transitRouteDisplay;
+  auto numMwmIds = make_shared<NumMwmIds>();
+  if (m_currentRouterType == RouterType::Transit)
+  {
+    m_delegate.RegisterCountryFilesOnRoute(numMwmIds);
+    auto getMwmId = [this, &numMwmIds](routing::NumMwmId numMwmId)
+    {
+      return m_callbacks.m_indexGetter().GetMwmIdByCountryFile(numMwmIds->GetFile(numMwmId));
+    };
+    transitRouteDisplay = make_unique<TransitRouteDisplay>(m_transitReadManager, getMwmId,
+                                                           m_callbacks.m_stringsBundleGetter,
+                                                           m_bmManager, m_transitSymbolSizes);
+  }
+
   vector<RouteSegment> segments;
   vector<m2::PointD> points;
   double distance = 0.0;
-  TransitRouteInfo transitRouteInfo;
+
   auto const subroutesCount = route.GetSubrouteCount();
   for (size_t subrouteIndex = route.GetCurrentSubrouteIdx(); subrouteIndex < subroutesCount; ++subrouteIndex)
   {
@@ -1057,35 +507,13 @@ void RoutingManager::InsertRoute(Route const & route)
       case RouterType::Transit:
         {
           subroute->m_routeType = df::RouteType::Transit;
-
-          auto numMwmIds = make_shared<NumMwmIds>();
-          m_delegate.RegisterCountryFilesOnRoute(numMwmIds);
-          auto getMwmIdFn = [this, &numMwmIds](routing::NumMwmId numMwmId)
-          {
-            return m_callbacks.m_indexGetter().GetMwmIdByCountryFile(numMwmIds->GetFile(numMwmId));
-          };
-
-          vector<TransitMarkInfo> transitMarks;
-          if (subrouteIndex > 0)
-          {
-            TransitStepInfo step;
-            step.m_type = TransitType::IntermediatePoint;
-            step.m_intermediateIndex = static_cast<int>(subrouteIndex - 1);
-            transitRouteInfo.AddStep(step);
-          }
-
-          FillTransitStyleForRendering(segments, m_transitReadManager, getMwmIdFn, m_callbacks.m_stringsBundleGetter,
-                                       *subroute.get(), transitMarks, transitRouteInfo);
+          transitRouteDisplay->ProcessSubroute(segments, *subroute.get());
 
           if (subroute->m_polyline.GetSize() < 2)
           {
             LOG(LWARNING, ("Invalid transit subroute. Points number =", subroute->m_polyline.GetSize()));
             continue;
           }
-
-          auto & marksController = m_bmManager->GetUserMarksController(UserMark::Type::TRANSIT);
-          CreateTransitMarks(transitMarks, m_transitSymbolSizes, marksController);
-          marksController.NotifyChanges();
           break;
         }
       case RouterType::Pedestrian:
@@ -1106,12 +534,13 @@ void RoutingManager::InsertRoute(Route const & route)
 
     // TODO: we will send subrouteId to routing subsystem when we can partly update route.
     //route.SetSubrouteUid(subrouteIndex, static_cast<SubrouteUid>(subrouteId));
-
     lock_guard<mutex> lock(m_drapeSubroutesMutex);
     m_drapeSubroutes.push_back(subrouteId);
-    transitRouteInfo.UpdateDistanceStrings();
-    m_transitRouteInfo = transitRouteInfo;
   }
+
+  lock_guard<mutex> lock(m_drapeSubroutesMutex);
+  m_transitRouteInfo = m_currentRouterType == RouterType::Transit ? transitRouteDisplay->GetRouteInfo()
+                                                                  : TransitRouteInfo();
 }
 
 void RoutingManager::FollowRoute()
